+++ extensions/api/data-management/transferprocess/src/main/java/org/eclipse/dataspaceconnector/api/datamanagement/transferprocess/TransferProcessApi.java
@@ -34,5 +34,5 @@ public interface TransferProcessApi {
 
     void deprovisionTransferProcess(String id);
 
-    String initiateTransfer(String assetId, TransferRequestDto transferRequest);
+    String initiateTransfer(TransferRequestDto transferRequest);
 }